gtuber: Do not use guint64 where unneeded
Streaming a video that has a bitrate over 4GB/s would be hard. Also I cannot imagine anyone sitting through a 136 years long movie ;-)

+++ gtuber/gtuber-media-info-devel.h
@@ -36,7 +36,7 @@ void              gtuber_media_info_set_title               (GtuberMediaInfo *in
 
 void              gtuber_media_info_set_description         (GtuberMediaInfo *info, const gchar *description);
 
-void              gtuber_media_info_set_duration            (GtuberMediaInfo *info, guint64 duration);
+void              gtuber_media_info_set_duration            (GtuberMediaInfo *info, guint duration);
 
 void              gtuber_media_info_add_stream              (GtuberMediaInfo *info, GtuberStream *stream);
